+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## V0.7.3
-- Added: Setting to customize rotation delay
-- Added: Setting to customize rotation amount
+- Added: Setting to customize snap rotation delay
+- Added: Setting to customize snap rotation amount
+- Added: Setting to customize smooth rotation speed
 - Fixed: Laser chronometer (Espionage DLC) no longer holsters instead of firing when triggered. Credit to Shoshana for the fix.
 
 ## V0.7.2